# optimize sketch reset